requirements.txt
pyspark
kafka-python
faker
pymysql
python-dotenv
requests
apache-airflow
